+++ extensions/arc/src/types.ts
@@ -31,6 +31,7 @@ export type NewTabSearchConfigs = {
 export type SearchConfig = {
   search: string;
   suggestions: string | null;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   suggestionParser: ((json: any, suggestions: Suggestion[]) => void) | null;
 };
 
@@ -52,3 +53,18 @@ export type Suggestion = {
 export type URLArguments = {
   url?: string;
 };
+
+/** Suggestion Parsers */
+interface GoogleClientData {
+  bpc: boolean;
+  tlw: boolean;
+}
+
+interface GoogleSuggestItem {
+  "google:clientdata": GoogleClientData;
+  "google:suggesttype": string[];
+  "google:verbatimrelevance": number;
+}
+
+export type GoogleSuggestionParser = [string, string[], string[], string[], GoogleSuggestItem];
+export type EcosiaSuggestionParser = [string, string[]];