+++ extensions/brew/src/preferences.ts
@@ -1,3 +1,3 @@
 import { getPreferenceValues } from "@raycast/api";
 
-export const preferences = <Preferences>getPreferenceValues();
+export const preferences = <Preferences & Preferences.CleanUp>getPreferenceValues();