+++ extensions/amazon-search/CHANGELOG.md
@@ -1,9 +1,23 @@
 # Amazon Search Changelog
 
+## [Enhancements] - 2024-02-25
+
+### Added
+
+- Added recent search history.
+
+### Changed
+
+- Improved empty state layout.
+
 ## [Enhancement] - 2023-12-30
+
+### Added
 
 - List section with suggestions count.
 
 ## [Initial Version] - 2023-12-02
+
+### Added
 
 - Initial version code.